+++ SDKChallengeProject/emojibattle/android/app/src/main/res/layout/activity_main.xml
@@ -1,25 +1,30 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".MainActivity">
 
-    <FrameLayout
-        android:id="@+id/localSurfaceView"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
+    <EditText
+        android:id="@+id/editChannel"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:hint="@string/create_channel"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        app:layout_constraintTop_toTopOf="parent"
+        tools:ignore="TextFields,AutoFill" />
 
-    <FrameLayout
-        android:id="@+id/remoteSurfaceView"
-        android:layout_width="180dp"
-        android:layout_height="300dp"
-        android:layout_alignParentEnd="true"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-</RelativeLayout>
+    <Button
+        android:id="@+id/btnBattle"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/enter_room"
+        app:layout_constraintEnd_toEndOf="@id/editChannel"
+        app:layout_constraintStart_toStartOf="@+id/editChannel"
+        app:layout_constraintTop_toBottomOf="@+id/editChannel" />
+
+
+</androidx.constraintlayout.widget.ConstraintLayout>
