+++ SDKChallengeProject/emojibattle/android/app/src/main/res/values/strings.xml
@@ -1,3 +1,7 @@
 <resources>
     <string name="app_name">emojibattle</string>
+    <string name="create_channel">创建房间</string>
+    <string name="enter_room">进入房间</string>
+    <string name="permission_dialog_title">Battle权限</string>
+    <string name="permission_dialog_message">需要相应的权限才能发起battle</string>
 </resources>
